# docs/swagger.yaml
basePath: /v1
definitions:
  api.ErrorDetail:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  api.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/api.ErrorDetail'
      metadata:
        $ref: '#/definitions/api.ResponseMetadata'
      success:
        type: boolean
    type: object
...
  api.ResponseMetadata:
    properties:
      timestamp:
        type: string
      trace_id:
        type: string
    type: object
  api.SuccessResponse:
    properties:
      data: {}
      metadata:
        $ref: '#/definitions/api.ResponseMetadata'
      success:
        type: boolean
    type: object
...
  services.TradingLogQueryResponse:
    properties:
      has_more:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
      trading_logs:
        items:
          $ref: '#/definitions/services.TradingLogResponse'
        type: array
    type: object
  services.TradingLogResponse:
    properties:
      id:
        type: string
      info:
        additionalProperties: true
        type: object
      message:
        type: string
      source:
        type: string
      sub_account_id:
        type: string
      timestamp:
        type: string
      trading_id:
        type: string
      transaction_id:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@tiris.ai
    name: Tiris API Support
    url: https://tiris.ai/support
  description: A Go-based microservice for quantitative trading data management, providing
    RESTful APIs for user management, trading integration, and trading operations.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://tiris.ai/terms
  title: Tiris Backend API
  version: "1.0"
paths:
  /admin/trading-logs:
    get:
      description: Lists all trading logs with filtering and pagination (admin only)
      parameters:
      - description: Filter by log type
        in: query
        name: type
        type: string
      - description: Filter by source
        enum:
        - manual
        - bot
        in: query
        name: source
        type: string
      - description: Start date (RFC3339 format)
        in: query
        name: start_date
        type: string
      - description: End date (RFC3339 format)
        in: query
        name: end_date
        type: string
      - default: 100
        description: Number of logs to return
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of logs to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.TradingLogQueryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all trading logs
      tags:
      - TradingLogs
  /admin/trading-logs/{id}:
    get:
      description: Retrieves a trading log by ID (admin only)
      parameters:
      - description: Trading Log ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.TradingLogResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get trading log by ID
      tags:
      - TradingLogs
...
  /trading-logs:
    get:
      description: Retrieves trading log history for the authenticated user with filtering
        and pagination
      parameters:
      - description: Filter by log type
        in: query
        name: type
        type: string
      - description: Filter by source
        enum:
        - manual
        - bot
        in: query
        name: source
        type: string
      - description: Start date (RFC3339 format)
        in: query
        name: start_date
        type: string
      - description: End date (RFC3339 format)
        in: query
        name: end_date
        type: string
      - default: 100
        description: Number of logs to return
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of logs to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.TradingLogQueryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user trading logs
      tags:
      - TradingLogs
    post:
      consumes:
      - application/json
      description: |-
        Creates a new trading log entry for the authenticated user.

        **Important**: The 'info' field structure must match the 'type' field:

        **Business Logic Types** (trigger automatic financial calculations):

        **For long/short/stop_loss types** - Required fields in 'info':
        - stock_account_id (string): Sub-account UUID for the asset (e.g., ETH account)
        - currency_account_id (string): Sub-account UUID for the currency (e.g., USDT account)
        - price (number): Price per unit (must be positive, up to 8 decimal places)
        - volume (number): Quantity traded (must be positive, up to 8 decimal places)
        - stock (string): Asset symbol, 1-20 characters (e.g., "ETH")
        - currency (string): Currency symbol, 1-20 characters (e.g., "USDT")
        - fee (number): Trading fee (must be non-negative, up to 8 decimal places)

        **For deposit/withdraw types** - Required fields in 'info':
        - account_id (string): Target sub-account UUID for the operation
        - amount (number): Amount to deposit/withdraw (must be positive, up to 8 decimal places)
        - currency (string): Currency symbol, 1-20 characters (e.g., "USDT")

        **Request Examples**:

        **Long Position Example:**
        <pre><code>{
        ⠀⠀"trading_id": "453f0347-3959-49de-8e3f-1cf7c8e0827c",
        ⠀⠀"type": "long",
        ⠀⠀"source": "bot",
        ⠀⠀"message": "ETH long position opened",
        ⠀⠀"info": {
        ⠀⠀⠀⠀"stock_account_id": "eth-account-uuid",
        ⠀⠀⠀⠀"currency_account_id": "usdt-account-uuid",
        ⠀⠀⠀⠀"price": 3000.00,
        ⠀⠀⠀⠀"volume": 2.0,
        ⠀⠀⠀⠀"stock": "ETH",
        ⠀⠀⠀⠀"currency": "USDT",
        ⠀⠀⠀⠀"fee": 12.00
        ⠀⠀}
        }</code></pre>

        **Deposit Example:**
        <pre><code>{
        ⠀⠀"trading_id": "453f0347-3959-49de-8e3f-1cf7c8e0827c",
        ⠀⠀"type": "deposit",
        ⠀⠀"source": "api",
        ⠀⠀"message": "USDT deposit to account",
        ⠀⠀"info": {
        ⠀⠀⠀⠀"account_id": "usdt-account-uuid",
        ⠀⠀⠀⠀"amount": 1000.00,
        ⠀⠀⠀⠀"currency": "USDT"
        ⠀⠀}
        }</code></pre>

        **Other Types**: Can use any object structure in the 'info' field
      parameters:
      - description: Create trading log request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.CreateTradingLogRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/services.TradingLogResponse'
        "400":
          description: 'Bad Request - Invalid request format, missing required fields,
            or incorrect ''info'' structure for the specified ''type''. Common validation
            errors: Missing required ''info'' fields for business logic types, Invalid
            data types or values in ''info'' fields, Non-existent sub-account IDs
            referenced in ''info'' fields'
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found - Trading ID or sub-account IDs referenced
            in 'info' field do not exist
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Unprocessable Entity - Business logic validation failed (e.g.,
            insufficient balance for withdraw operations)
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create trading log
      tags:
      - TradingLogs
  /trading-logs/{id}:
    delete:
      description: Deletes a trading log entry (must belong to authenticated user
        and be manual)
      parameters:
      - description: Trading Log ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete trading log
      tags:
      - TradingLogs
    get:
      description: Retrieves a specific trading log by ID (must belong to authenticated
        user)
      parameters:
      - description: Trading Log ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.TradingLogResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get trading log by ID
      tags:
      - TradingLogs
...
securityDefinitions:
  BearerAuth:
    description: 'Enter JWT Bearer token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...